+++ tests/common/widgets/hits/insights.ts
@@ -26,7 +26,7 @@ export function createInsightsTests(
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = Object.assign(jest.fn(), { version: '2.15.0' });
+      window.aa = Object.assign(jest.fn(), { version: '2.17.2' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
@@ -138,7 +138,7 @@ export function createInsightsTests(
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 25;
-      window.aa = Object.assign(jest.fn(), { version: '2.15.0' });
+      window.aa = Object.assign(jest.fn(), { version: '2.17.2' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
@@ -274,7 +274,7 @@ export function createInsightsTests(
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = Object.assign(jest.fn(), { version: '2.15.0' });
+      window.aa = Object.assign(jest.fn(), { version: '2.17.2' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
@@ -356,7 +356,7 @@ export function createInsightsTests(
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = Object.assign(jest.fn(), { version: '2.15.0' });
+      window.aa = Object.assign(jest.fn(), { version: '2.17.2' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
@@ -445,7 +445,7 @@ export function createInsightsTests(
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = Object.assign(jest.fn(), { version: '2.15.0' });
+      window.aa = Object.assign(jest.fn(), { version: '2.17.2' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
@@ -531,7 +531,7 @@ export function createInsightsTests(
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = Object.assign(jest.fn(), { version: '2.15.0' });
+      window.aa = Object.assign(jest.fn(), { version: '2.17.2' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
@@ -628,7 +628,7 @@ export function createInsightsTests(
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = Object.assign(jest.fn(), { version: '2.15.0' });
+      window.aa = Object.assign(jest.fn(), { version: '2.17.2' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',
@@ -710,7 +710,7 @@ export function createInsightsTests(
       const delay = 100;
       const margin = 10;
       const hitsPerPage = 2;
-      window.aa = Object.assign(jest.fn(), { version: '2.15.0' });
+      window.aa = Object.assign(jest.fn(), { version: '2.17.2' });
       const options = {
         instantSearchOptions: {
           indexName: 'indexName',